Create scrape_GAO_CRA.R
This is not exactly a scraper. Rather, it is a simple file that pulls the universe of rules submitted under the Congressional Review Act. Note that approximately 90% of rules are reported (compared to a much lower reporting rate to the Unified Agenda). Additionally, this file will pull guidance documents; it is difficult to find additional information on them. Also note that some of the sub-agencies listed are not accurate and some organizational units reported to the GAO are not listed under agency statute or organizational charts. What does this mean? Well, you should probably hand-code the data to find the errors or contact me for the cleaned data.